requirements.txt
asgiref==3.11.1
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.1.2
Django==4.2.11
django-ckeditor==6.7.3
django-js-asset==3.1.2
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
webencodings==0.5.1
whitenoise==6.12.0
